Ajustando json de retorno

+++ app/src/main/kotlin/com/trmamobilesolutions/alertcoin/home/model/domain/Total.kt
@@ -1,6 +1,6 @@
 package com.trmamobilesolutions.alertcoin.home.model.domain
 
-data class Total(val high: Int = 0,
+data class Total(val high: Double = 0.0,
                  val vol: Double = 0.0,
                  val last: Double = 0.0,
                  val money: Double = 0.0,
